+++ IDP - Dynamic Mapping/Assets/Scripts/KalmanFilter/StateCovariance.cs
@@ -10,9 +10,9 @@ public class StateCovariance {
      * Pvm(i|j): cross-covariance matrix between vehicle and landmark states
      */
 
-    private static MatrixBuilder<float> M = Matrix<float>.Build;
+    private static MatrixBuilder<double> M = Matrix<double>.Build;
 
-    private Matrix<float> P;
+    private Matrix<double> P;
     private static int STATE_DIM;
     private static int LANDMARK_DIM;
 
@@ -24,11 +24,11 @@ public class StateCovariance {
         P = M.DenseDiagonal(size, size, 0);
     }
 
-    private StateCovariance(Matrix<float> P) {
+    private StateCovariance(Matrix<double> P) {
         this.P = P;
     }
 
-    private StateCovariance(Matrix<float> Pvv, Matrix<float> Pvm, Matrix<float> Pmm) {
+    private StateCovariance(Matrix<double> Pvv, Matrix<double> Pvm, Matrix<double> Pmm) {
         P = M.Dense(Pvv.RowCount + Pmm.RowCount, Pvv.ColumnCount + Pmm.ColumnCount);
         P.SetSubMatrix(0, 0, Pvv);
         P.SetSubMatrix(0, STATE_DIM, Pvm);
@@ -36,29 +36,29 @@ public class StateCovariance {
         P.SetSubMatrix(STATE_DIM, STATE_DIM, Pmm);
     }
 
-    public Matrix<float> ExtractPvv() {
+    public Matrix<double> ExtractPvv() {
         return P.SubMatrix(0, STATE_DIM, 0, STATE_DIM);
     }
 
-    public Matrix<float> ExtractLandmarkCovariance(int landmarkIndex) {
+    public Matrix<double> ExtractLandmarkCovariance(int landmarkIndex) {
         int index = STATE_DIM + landmarkIndex * LANDMARK_DIM;
         return P.SubMatrix(index, LANDMARK_DIM, index, LANDMARK_DIM);
     }
 
-    public void AddLandmark(Matrix<float> landmarkCovariance) {
-        Matrix<float> newP = M.Dense(P.RowCount+LANDMARK_DIM, P.ColumnCount+LANDMARK_DIM);
+    public void AddLandmark(Matrix<double> landmarkCovariance) {
+        Matrix<double> newP = M.Dense(P.RowCount+LANDMARK_DIM, P.ColumnCount+LANDMARK_DIM);
         newP.SetSubMatrix(0, 0, P);
         newP.SetSubMatrix(P.RowCount, P.ColumnCount, landmarkCovariance);
 
         P = newP;
     }
 
-    public StateCovariance PredictStateEstimateCovariance(Matrix<float> Fv, VehicleModel model) {
+    public StateCovariance PredictStateEstimateCovariance(Matrix<double> Fv, VehicleModel model) {
         int landmarksSize = P.ColumnCount - STATE_DIM;
 
-        Matrix<float> Pvv = P.SubMatrix(0, STATE_DIM, 0, STATE_DIM);
-        Matrix<float> Pvm = P.SubMatrix(0, STATE_DIM, STATE_DIM, landmarksSize);
-        Matrix<float> Pmm = P.SubMatrix(STATE_DIM, landmarksSize, STATE_DIM, landmarksSize);
+        Matrix<double> Pvv = P.SubMatrix(0, STATE_DIM, 0, STATE_DIM);
+        Matrix<double> Pvm = P.SubMatrix(0, STATE_DIM, STATE_DIM, landmarksSize);
+        Matrix<double> Pmm = P.SubMatrix(STATE_DIM, landmarksSize, STATE_DIM, landmarksSize);
 
         return new StateCovariance(
             Fv * Pvv.TransposeAndMultiply(Fv) + model.ProcessNoiseError,
@@ -66,20 +66,20 @@ public class StateCovariance {
             Pmm);
     }
 
-    public (Matrix<float>, Matrix<float>) ComputeInnovationAndGainMatrices(Matrix<float> H, Matrix<float> R) {
-        Matrix<float> tmp = P.TransposeAndMultiply(H);
+    public (Matrix<double>, Matrix<double>) ComputeInnovationAndGainMatrices(Matrix<double> H, Matrix<double> R) {
+        Matrix<double> tmp = P.TransposeAndMultiply(H);
 
-        Matrix<float> Si = H * tmp + R;           // Matrix(OBSERVATION_DIM, OBSERVATION_DIM)
-        Matrix<float> Wi = tmp * Si.Inverse();    // Matrix(STATE_DIM+landmarkCount*LANDMARK_DIM, OBSERVATION_DIM)
+        Matrix<double> Si = H * tmp + R;           // Matrix(OBSERVATION_DIM, OBSERVATION_DIM)
+        Matrix<double> Wi = tmp * Si.Inverse();    // Matrix(STATE_DIM+landmarkCount*LANDMARK_DIM, OBSERVATION_DIM)
 
         return (Si, Wi);
     }
 
-    public static StateCovariance operator +(StateCovariance state, Matrix<float> other) {
+    public static StateCovariance operator +(StateCovariance state, Matrix<double> other) {
         return new StateCovariance(state.P + other);
     }
 
-    public static StateCovariance operator -(StateCovariance state, Matrix<float> other) {
+    public static StateCovariance operator -(StateCovariance state, Matrix<double> other) {
         return new StateCovariance(state.P - other);
     }
 }
